Addressing latest comments
