+++ src/main/java/io/github/mdcdxcvi/EXPLORER.java
@@ -1,4 +1,4 @@
-package com.mdcdxcvi;
+package io.github.mdcdxcvi;
 
 import java.io.*;
 public class EXPLORER{
