+++ src/main/java/io/github/mdcdxcvi/GameScreenCli.java
@@ -1,4 +1,4 @@
-package com.mdcdxcvi;
+package io.github.mdcdxcvi;
 
 public class GameScreenCli {
     protected static void printThanksForPlayingScreen() {
